fix: iinsert fix for an user from same pg not view yours virtual key
