+++ packages/pn-personagiuridica-webapp/src/redux/apikeys/reducers.ts
@@ -1,12 +1,13 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-import { BffPublicKeysResponse, BffVirtualKeysResponse } from '../../generated-client/pg-apikeys';
-import { getPublicKeys, getVirtualApiKeys } from './actions';
+import { BffPublicKeysCheckIssuerResponse, BffPublicKeysResponse, BffVirtualKeysResponse, PublicKeysIssuerResponseIssuerStatusEnum } from '../../generated-client/pg-apikeys';
+import { getPublicKeys, getVirtualApiKeys, checkPublicKeyIssuer } from './actions';
 
 type initialStateType = {
   loading: boolean;
   publicKeys: BffPublicKeysResponse;
   virtualKeys: BffVirtualKeysResponse;
+  issuerState: BffPublicKeysCheckIssuerResponse;
 };
 
 const initialState: initialStateType = {
@@ -19,6 +20,13 @@ const initialState: initialStateType = {
     items: [],
     total: 0,
   },
+  issuerState:{
+    tosAccepted: false,
+    issuer: {
+      isPresent: false,
+      issuerStatus: PublicKeysIssuerResponseIssuerStatusEnum.Inactive,
+    }
+  }
 };
 
 /* eslint-disable functional/immutable-data */
@@ -31,6 +39,9 @@ const apiKeysSlice = createSlice({
   extraReducers: (builder) => {
     builder.addCase(getPublicKeys.fulfilled, (state, action) => {
       state.publicKeys = action.payload;
+    });
+    builder.addCase(checkPublicKeyIssuer.fulfilled, (state, action) => {
+      state.issuerState = action.payload;
     });
     builder.addCase(getVirtualApiKeys.fulfilled, (state, action) => {
       state.virtualKeys = action.payload;
